refactor(core): consolidar soft delete en core/soft_delete
- Migra imports legacy core.soft_delete_* a core.soft_delete.*

- Elimina wrappers de compatibilidad en core/

- Mantiene comportamiento funcional sin cambios
